+++ extensions/easydict/src/components.tsx
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-06-26 11:13
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-21 15:40
+ * @lastEditTime: 2022-07-23 13:28
  * @fileName: components.tsx
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -23,6 +23,7 @@ import {
   QueryWordInfo,
   SectionType,
   TranslationType,
+  WebTranslationItem,
   YoudaoTranslateReformatResultItem,
 } from "./types";
 import {
@@ -275,13 +276,6 @@ export function getWordAccessories(
   return wordAccessories;
 }
 
-interface WebTranslationItem {
-  type: QueryType;
-  webUrl: string;
-  icon: Image.ImageLike;
-  title: string;
-}
-
 /**
  * Return WebTranslationItem according to the query type and info
  */
@@ -313,5 +307,6 @@ export function getWebTranslationItem(
       break;
     }
   }
+  // console.log("---> getWebTranslationItem", webUrl);
   return webUrl ? { type: queryType, webUrl, icon, title } : undefined;
 }