+++ extensions/easydict/src/releaseVersion/changelog.ts
@@ -2,19 +2,20 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-21 16:03
+ * @lastEditTime: 2022-07-23 18:25
  * @fileName: changelog.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const changelog = `
-## [v1.3.1] - 2022-07-21
+## [v1.4.0] - 2022-07-23
 
-### 🐞 Fixes
+### ✨ 新特性
 
-- 修复了可能多次显示更新日志提示的问题。
-- 修复了有道翻译请求异常错误未处理问题。
+- 支持 Google 翻译。
+
+![Google](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/easydict-6-1658571181.png)
 
 ### ✅ 最近更新
 
@@ -22,16 +23,12 @@ export const changelog = `
 - 支持手动排序翻译结果显示。
 - 支持苹果🍎系统翻译
 - 修复了删除输入框文本会导致重复查询选中文本问题。
-- 修复了动作按钮渲染时会闪烁显示 “✨ New Version Released” 问题。
-
-![easydict-5](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/easydict-5-1658309785.png)
 
 ---
 
-### 🐞 Fixes
+### ✨ New
 
-- fixed new release prompt may be shown multiple times.
-- handled the exceptional error of request to Youdao translation.
+- Added support for Google translation.
 
 ### ✅ Recent Updates
 
@@ -39,5 +36,4 @@ export const changelog = `
 - Supported to sort translation results manually.
 - Added support for Apple 🍎 system translation.
 - Fixed deleting input box text will cause repeated query selected text problem.
-- Fixed the bug of "✨ New Version Released" flashing when rendering action button.
 `;